requirements.txt
Pillow == 9.2.0
PyQRCode == 1.2.1
pypng == 0.20220715.0